third_party/blink/web_tests/platform/mac/external/wpt/editing/run/strikethrough_2001-last-expected.txt

This is a testharness.js-based test.
Found 22 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<s>[bar]</s>baz" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<s>[bar]</s>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<s>[bar]</s>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<s>[bar]</s>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo{<s>bar</s>}baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo{<s>bar</s>}baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo{<s>bar</s>}baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo{<s>bar</s>}baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""]] "fo[o<span style=text-decoration:line-through>b]ar</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<strike>o</strike><s>bar</s>baz" but got "fo<strike>ob</strike><span style=\\"text-decoration:line-through\\">ar</span>baz"
[FAIL] [["strikethrough",""]] "fo[o<span style=text-decoration:line-through>b]ar</span>baz" queryCommandIndeterm("strikethrough") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""]] "fo[o<span style=text-decoration:line-through>b]ar</span>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""]] "fo[o<span style=text-decoration:line-through>b]ar</span>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>fo[o</strike><s>b]ar</s>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>fo[o</strike><s>b]ar</s>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<strike>fo[o</strike><s>b]ar</s>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<strike>fo[o</strike><s>b]ar</s>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>fo[o</s><del>b]ar</del>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>fo[o</s><del>b]ar</del>" queryCommandIndeterm("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>fo[o</s><del>b]ar</del>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>fo[o</s><del>b]ar</del>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>fo[o</s><del>b]ar</del>" queryCommandIndeterm("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>fo[o</s><del>b]ar</del>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
Harness: the test ran to completion.
